Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -1,20 +1,20 @@
 #include <stdio.h>
 int main() {
-    float a,b;
+    int a,b;
     char c;
-    scanf("%f %f",&a,&b);
+    scanf("%d %d",&a,&b);
     scanf(" %c",&c);
     if(c=='+'){
-        printf("%f",a+b);
+        printf("%d",a+b);
     }
     else if(c=='-'){
-        printf("%f",a-b);
+        printf("%d",a-b);
     }
     else if(c=='*'){
-        printf("%f",a*b);
+        printf("%d",a*b);
     }
     else if(c=='/'){
-        printf("%f",a/b);
+        printf("%d",a/b);
     }
     return 0;
 }
